Pilotscontroller UI
- can now list/details/add/update and delete pilots
- made pilotsmodel so i can have a dropdown of sex and can have length in feet and wheight in lbs
-experimented with img

+++ RacingApp.Core/DTO'S/PilotsDTO.cs
@@ -17,6 +17,6 @@ namespace RacingApp.Core.DTO_S
         public char Sex { get; set; }
         public DateTime DateOfBirth { get; set; }
         public int Length { get; set; }
-        public decimal Weight { get; set; }
+        public double Weight { get; set; }
     }
 }
